5 commit

+++ src/main/java/DAO/EmployeeDAO.java
@@ -6,7 +6,7 @@ import model.Employee;
 import java.util.List;
 
 public interface EmployeeDAO {
-    Employee findById(Long id);
+    Employee findById(Integer id);
     void update(Employee employee);
     void save(Employee employee);
     void delete(Employee employee);
